+++ synehr/data_gen.py
@@ -1,9 +1,33 @@
 from auxillary import genDOB
 
 from auxillary import fake_factory
-
+import numpy as np
 import pandas as pd
 import numpy.random as npr
+import random
+
+def gen_typo(Data):
+
+    arr = np.array(Data['first'])
+    characters = 'qwertyuioplkjhgfdsazxcvbnm'
+
+    for i in range(0, len(arr)):
+        x = arr[i]
+        # print x
+        # print len(x)
+        if len(x) > 3:
+            if (random.randint(0, 2) == 1):
+                rnd = random.randint(2, len(x) - 2)
+                # print rnd
+                tmp1 = random.randint(0, len(characters))
+                rndCharacter = characters[tmp1:tmp1 + 1]
+                # print rndCharacter
+                # x[rnd:rnd+1] = rndCharacter
+                x = x[0:rnd] + rndCharacter + x[rnd + 1:]
+                arr[i] = x
+    Data['first']=arr
+    print "Induced Errors: ",Data['first']
+    return Data['first']
 
 def data_gen(first_names_data,last_names_data,size,race_ratio,male_gender):
     ### sizes for each race
@@ -32,9 +56,9 @@ def data_gen(first_names_data,last_names_data,size,race_ratio,male_gender):
     #print spanish_male_size
     spanish_female_size = int(spanish_size * female_gender)
     afr_amer_male_size = int(afr_amer_size * male_gender)
-    print "afr_amer_male_size: ",afr_amer_male_size
+    #print "afr_amer_male_size: ",afr_amer_male_size
     afr_amer_female_size = int(afr_amer_size * female_gender)
-    print "afr_amer_female_size: ",afr_amer_female_size
+    #print "afr_amer_female_size: ",afr_amer_female_size
     caucasian_male_size = int(caucasian_size * male_gender)
     caucasian_female_size = int(caucasian_size * female_gender)
     #print caucasian_male_size
@@ -87,7 +111,7 @@ def data_gen(first_names_data,last_names_data,size,race_ratio,male_gender):
     asian_indian_fn = asian_data_fn.loc[(asian_data_fn['detailed_race'] == 'Indian')]
     # Check how many records are there
     indian_rows = len(asian_indian_fn)
-    print "no of rows: ",indian_rows
+    #print "no of rows: ",indian_rows
     # Extract Chinese Last Names
     asian_indian_ln = asian_dataset_ln.loc[(asian_dataset_ln['detailed_race'] == 'Indian')]
     # Select as many Last names randomly as you need
@@ -120,18 +144,18 @@ def data_gen(first_names_data,last_names_data,size,race_ratio,male_gender):
 
 
     afr_amer_male_fn = first_names_data.loc[(first_names_data['race'] == 'African_American') & (first_names_data['gender'] == 'M'),]
-    print "afro male:", len(afr_amer_male_fn)
+    #print "afro male:", len(afr_amer_male_fn)
     afr_amer_female_fn = first_names_data.loc[(first_names_data['race'] == 'African_American') & (first_names_data['gender'] == 'F'),]
-    print "afro female: ",len(afr_amer_female_fn)
+    #print "afro female: ",len(afr_amer_female_fn)
     afr_amer_male_data=afr_amer_male_fn.sample(n=afr_amer_male_size,replace=True)
-    print len(afr_amer_male_data)
+    #print len(afr_amer_male_data)
     afr_amer_female_data=afr_amer_female_fn.sample(n=afr_amer_female_size,replace=True)
-    print len(afr_amer_female_data)
+    #print len(afr_amer_female_data)
     afr_amer_data_fn=pd.concat([afr_amer_male_data,afr_amer_female_data])
 
     afr_amer_ln=afr_amer_dataset_ln.sample(n=(afr_amer_male_size+afr_amer_female_size), replace=True)
     afr_amer_data_fn.insert(1,'last',afr_amer_ln['last'].tolist())
-    print "African American Head: ",afr_amer_female_fn.head()
+    #print "African American Head: ",afr_amer_female_fn.head()
 
 
 
@@ -140,14 +164,14 @@ def data_gen(first_names_data,last_names_data,size,race_ratio,male_gender):
     native_amer_male_data = native_amer_male_fn.sample(n=native_amer_male_size, replace=True)
     native_amer_female_data = native_amer_female_fn.sample(n=native_amer_female_size, replace=True)
     native_amer_data_fn = pd.concat([native_amer_male_data, native_amer_female_data])
-    print "nat fn len: ",len(native_amer_data_fn)
+    #print "nat fn len: ",len(native_amer_data_fn)
 
     native_ln = native_amer_alaskan_dataset_ln.sample(n=(native_amer_male_size + native_amer_female_size), replace=True)
-    print native_ln.head()
-    print "Length of Nat last names:",len(native_ln)
+    #print native_ln.head()
+    #print "Length of Nat last names:",len(native_ln)
     native_amer_data_fn.insert(1, 'last', native_ln['last'].tolist())
 
-    print native_amer_data_fn.head()
+    #print native_amer_data_fn.head()
 
     caucasian_male_fn = first_names_data.loc[(first_names_data['race'] == 'Caucasian') & (first_names_data['gender'] == 'M'),]
     caucasian_female_fn = first_names_data.loc[(first_names_data['race'] == 'Caucasian') & (first_names_data['gender'] == 'F'),]
@@ -158,7 +182,7 @@ def data_gen(first_names_data,last_names_data,size,race_ratio,male_gender):
     cauc_ln = caucasian_dataset_ln.sample(n=(caucasian_male_size + caucasian_female_size), replace=True)
     caucasian_data_fn.insert(1, 'last', cauc_ln['last'].tolist())
 
-    mixed_dataset = first_names_data.sample(n=mixed_size * 1000, replace=True)
+    mixed_dataset = first_names_data
     mixed_dataset['race'] = "Mixed"
     mixed_dataset['detailed_race']="None"
 
@@ -191,10 +215,9 @@ def mk_data(first_names_data, last_names_data, min_date, max_date, size,male_gen
     # dataset list by race
 
     # data generator by race
-    #print first_names_data.head()
-    #print last_names_data.head()
-    #print size
-    #print race_ratio
+    print "Calling the data generator function..."
+    print size
+    print race_ratio
     race_generator = data_gen(first_names_data, last_names_data, size,race_ratio,male_gender)
 
     # fake factory
@@ -214,12 +237,12 @@ def mk_data(first_names_data, last_names_data, min_date, max_date, size,male_gen
     print "Size of data: ",len(data)
     data1=data.dropna(how='any')
     print "After dropping any with na: ",len(data1)
-    error_rate=0.1*size
-    error_index=npr.randint(0,len(data),size=error_rate)
-    data_err=data.ix[error_index]
+    error_rate=int(0.1*size)
+    error_index=npr.randint(0,len(data1),size=error_rate)
+    data_err=data1.ix[error_index]
     print "Data to generate Error: ",data_err.head()
-    #data_err=gen_typo(data_err)
-    #data.ix[error_index]=data_err
+    data_err=gen_typo(data_err)
+    data1.ix[error_index]=data_err
 
     return data
